항공을 제외한 모든 파일 병합 완료. 게시판 기능은 로그인과 연결 후 완벽 구현. 디자인만 손보면 됨. 마이페이지는 내가쓴글과 예약내역 부분이 전혀 안되어있음. 전체적으로 디자인은 다 손봐야함. 로그인이 좀 이상함 숙박도 거의 다됐는데 로그인이랑 연동부분 체크해야함 즐겨찾기, 예약 등 채팅도 로그인과 연결되면 손볼거 좀 있고, 디자인은 건드릴거 없음.

+++ android_studio/app/src/main/java/bitc/fullstack502/android_studio/network/dto/LodgingBookingDto.kt
@@ -5,7 +5,7 @@ data class LodgingBookingDto(
     val lodId: Long,
     val ckIn: String,
     val ckOut: String,
-    val totalPrice: Int,
+    val totalPrice: Long,
     val roomType: String,
     val adult: Int,
     val child: Int,
